Models/GGSNN-GMN/NeuralNetwork/requirements.txt
coloredlogs==15.0.1
dm-sonnet==1.34
MarkupSafe==2.0.0
networkx==2.5
pandas==1.3.5
scikit-learn==1.0.2
tensorflow-estimator==1.14.0
tensorflow-probability==0.7
tensorflow==1.14.0
tqdm==4.64.0